requirements.txt
mock
django>=1.3
coverage>=3.5.3
django-discover-runner>=0.2.2